adding 3rd party folders to git
adding start of grid movement based on previous grid movement in burgertime

+++ Minigin/Subject.cpp
@@ -18,7 +18,7 @@ void dae::Subject::RemoveObserver(std::shared_ptr<Observer> observer)
 
 void dae::Subject::Notify(std::shared_ptr<GameObject> gameObject, ObserverEvent event)
 {
-	for (int i {}; i < m_Observers.size(); ++i)
+	for (size_t i {}; i < m_Observers.size(); ++i)
 	{
 		m_Observers[i]->Notify(gameObject, event);
 	}
